requirements.txt
flask
torch
torchvision
Pillow
numpy
joblib
wikipedia
requests
